refactor: extract bridge drawing functions from CanvasIsometricGrid
Move drawBridgeTile, drawSuspensionBridgeTowers, and drawSuspensionBridgeOverlay
functions (~780 lines) to bridgeDrawing.ts to reduce the size of
CanvasIsometricGrid.tsx and improve maintainability.
